Updates for issue #432:
Abstracting architecture_029 to identifier_alignment_rule.
